+++ src/test/components/arrowFunctionDeclarationEmpty/arrowFunctionDeclarationEmpty.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function Component() {
+  return (
+    <div className='my-class-2'>
+      <div className='my-class'>Test</div>
+    </div>
+  );
+}